+++ utils/proxy.js
@@ -1,8 +1,0 @@
-//Internal utility for creating context style global utils
-const proxy = function(method) {
-  return function(context, ...args) {
-    return method.apply(context, args);
-  };
-};
-
-export default proxy;